requirements.txt
Django==4.2.9
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
openpyxl==3.1.2
python-decouple==3.8
gunicorn==21.2.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
